Types fixed.

+++ include/KGOptions.h
@@ -71,7 +71,7 @@ extern "C"
         bool shortOutput;   ///< Reduced output (generated key only)
         bool showHelp;  ///< Print the help
         bool showVersion;   ///< Print the version
-        int keyLength;      ///< Keylength
+        long keyLength;      ///< Keylength
         enum Format keyFormat;  ///< Key format
     };
